+++ src/gui.js
@@ -87,11 +87,11 @@ BlockVisibilityDialogMorph, ThreadManager, isString, SnapExtensions, snapEquals
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.gui = '2024-July-22';
+modules.gui = '2024-July-24';
 
 // Declarations
 
-var SnapVersion = '10-rc5';
+var SnapVersion = '10.0.0';
 
 var database = new Database()
 var IDE_Morph;
@@ -6061,6 +6061,7 @@ IDE_Morph.prototype.aboutSnap = function () {
         + '\nSimon Mong: Custom blocks palette arrangement'
         + '\nIvan Motyashov: Initial Squeak Porting'
         + '\nLucas Karahadian: Piano Keyboard Design'
+        + '\nego-lay-atman-bay: Piano Keyboard Octave Switching'
         + '\nDavide Della Casa: Morphic Optimizations'
         + '\nAchal Dave: Web Audio'
         + '\nJoe Otto: Morphic Testing and Debugging'