+++ HISTORY.md
@@ -1,6 +1,8 @@
 # Snap! (BYOB) History
 
 ## in development:
+
+## 10.0.0:
 * **New Features:**
     1. Blocks all the way
         * all blocks in the palette are now custom blocks, except hat blocks and rings
@@ -81,6 +83,7 @@
         * new "svg poly" export format for vector pen trails, optimized for speed and laser-cutting
         * export vector pen trails as embroidery files (DST, EXP)
         * added pen trails export options to the project ("file") menu
+        * added ability to switch octaves in the piano keyboard menu, thanks, ego-lay-atman-bay!
 * **Notable Changes:**
     1. Blocks all the way
         * renamed "primitive" blocks into "extension"
@@ -122,6 +125,14 @@
     * updated extensions.md
 * **Translation Updates:**
     * German
+
+### 2024-07-24
+* widgets: added ability to switch octaves in the piano keyboard widget, thanks, ego-lay-atman-bay!
+* gui: added credits for ego-lay-atman-bay
+* v10-rc6
+* reverted piano keyboard behavior
+* v10-rc7
+* prepared v10.0.0 release
 
 ### 2024-07-22
 * edited help screen for doDeleteBlock, converted to PNG, reduced file size